refactor(tui): own clipboard image acquisition
